+++ statistics-impl/src/main/java/com/dwarfeng/statistics/impl/service/telqos/DriveCommand.java
@@ -1,0 +1,75 @@
+package com.dwarfeng.statistics.impl.service.telqos;
+
+import com.dwarfeng.springtelqos.node.config.TelqosCommand;
+import com.dwarfeng.springtelqos.sdk.command.CliCommand;
+import com.dwarfeng.springtelqos.stack.command.Context;
+import com.dwarfeng.springtelqos.stack.exception.TelqosException;
+import com.dwarfeng.statistics.stack.service.DriveQosService;
+import org.apache.commons.cli.CommandLine;
+import org.apache.commons.cli.Option;
+import org.apache.commons.lang3.tuple.Pair;
+
+import java.util.ArrayList;
+import java.util.List;
+
+@TelqosCommand
+public class DriveCommand extends CliCommand {
+
+    private static final String COMMAND_OPTION_STATUS = "status";
+
+    private static final String[] COMMAND_OPTION_ARRAY = new String[]{
+            COMMAND_OPTION_STATUS
+    };
+
+    private static final String IDENTITY = "drive";
+    private static final String DESCRIPTION = "驱动处理器操作/查看";
+
+    private static final String CMD_LINE_SYNTAX_STATUS = IDENTITY + " " +
+            CommandUtil.concatOptionPrefix(COMMAND_OPTION_STATUS);
+
+    private static final String[] CMD_LINE_ARRAY = new String[]{
+            CMD_LINE_SYNTAX_STATUS
+    };
+
+    private static final String CMD_LINE_SYNTAX = CommandUtil.syntax(CMD_LINE_ARRAY);
+
+    private final DriveQosService driveQosService;
+
+    public DriveCommand(DriveQosService driveQosService) {
+        super(IDENTITY, DESCRIPTION, CMD_LINE_SYNTAX);
+        this.driveQosService = driveQosService;
+    }
+
+    @Override
+    protected List<Option> buildOptions() {
+        List<Option> list = new ArrayList<>();
+        list.add(Option.builder(COMMAND_OPTION_STATUS).desc("查看驱动处理器状态").build());
+        return list;
+    }
+
+    @SuppressWarnings("SwitchStatementWithTooFewBranches")
+    @Override
+    protected void executeWithCmd(Context context, CommandLine cmd) throws TelqosException {
+        try {
+            Pair<String, Integer> pair = CommandUtil.analyseCommand(cmd, COMMAND_OPTION_ARRAY);
+            if (pair.getRight() != 1) {
+                context.sendMessage(CommandUtil.optionMismatchMessage(COMMAND_OPTION_ARRAY));
+                context.sendMessage(CMD_LINE_SYNTAX);
+                return;
+            }
+            switch (pair.getLeft()) {
+                case COMMAND_OPTION_STATUS:
+                    printStatus(context);
+                    break;
+            }
+        } catch (Exception e) {
+            throw new TelqosException(e);
+        }
+    }
+
+    private void printStatus(Context context) throws Exception {
+        boolean startedFlag = driveQosService.isStarted();
+
+        context.sendMessage(String.format("started: %b.", startedFlag));
+    }
+}